slack-aggregate-threads/requirements.txt
quixstreams<2.6
python-dotenv